Adding replication interface.
git-svn-id: https://opendentalsoft.com:23793/svn/opendental/head@3010 098d0599-7f2b-0410-95a5-ad30ac85393a

+++ OpenDentBusiness/DataConnection.cs
@@ -239,7 +239,7 @@ namespace OpenDentBusiness{
 				cmd.CommandText="SELECT LAST_INSERT_ID()";
 				dr=(MySqlDataReader)cmd.ExecuteReader();
 				if(dr.Read()) {
-					InsertID=Convert.ToInt32(dr[0].ToString());
+					InsertID=Convert.ToInt64(dr[0].ToString());
 				}
 			}
 			con.Close();
